refactor: extract helper to read input users
